requirements.txt
tensorflow-gpu
matplotlib
opencv-python
